Changed int to unsigned long long int

+++ factorial.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
-int calculateFactorial(int x)
+unsigned long long int calculateFactorial(long long int x)
 {
-    int result = 1;
-    for (int i = x; i > 1; --i)
+    unsigned long long int result = 1;
+    for (unsigned long long int i = x; i > 1; --i)
     {
         result *= i;
     }
@@ -13,9 +13,9 @@ int calculateFactorial(int x)
 int main(void)
 {
     printf("Enter the number to calculate the factorial of: ");
-    int x;
-    scanf("%i", &x);
-    int xFactorial = calculateFactorial(x);
-    printf("%d! = %d\n", x, xFactorial);
+    unsigned long long int x;
+    scanf("%llu", &x);
+    unsigned long long int xFactorial = calculateFactorial(x);
+    printf("%d! = %llu\n", x, xFactorial);
     return 0;
 }
